[68-advanced-search] Search app renamed to advanced search
